#79 Rename simulator module
